User: refactor user avatar to its own controller
